requirements.txt
Django
scikit-learn
python-docx
PyMuPDF
Pillow
python-dotenv
WeasyPrint
google-generativeai
celery
redis
eventlet
django-crispy-forms
crispy-tailwind